Future/FTrader-Backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.1
django-allauth==0.60.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
python-dotenv==1.0.1
requests==2.31.0
websocket-client==1.7.0
gunicorn==21.2.0
whitenoise==6.6.0
pytest==7.4.4
pytest-django==4.7.0
coverage==7.4.1
flake8==7.0.0
black==24.1.1
isort==5.13.2
coreapi==2.3.3
pyyaml==6.0.1
uritemplate==4.1.1
Pillow==10.1.0
lightstreamer-client
influxdb-client==1.36.0
pika==1.3.2

# Machine Learning Libraries
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
tensorflow==2.13.0
joblib==1.3.1
